+++ src/Farkle/Builder/Precompiler/Loader.fs
@@ -7,9 +7,12 @@ namespace Farkle.Builder.Precompiler
 
 open Farkle.Builder
 open Farkle.Grammar
+open System
 open System.Reflection
 
 module DFB = DesigntimeFarkleBuild
+
+type PrecompilerLoaderException(msg, innerExn: exn) = inherit exn(msg, innerExn)
 
 /// A kind of designtime Farkle whose grammar can be precompiled
 /// when the assembly containing it is compiled. Its distinguishing
@@ -47,6 +50,19 @@ type private PrecompilableDesigntimeFarkle<'T> (df: DesigntimeFarkle, asm) =
 /// from assemblies using reflection.
 module internal Loader =
 
+    let private suppressLoaderErrorsSwitch = "Switch.Farkle.Builder.Precompiler.SuppressLoaderErrors"
+
+    let private shouldSuppressLoaderErrors<'a> =
+        let switchFound, switchEnabled = AppContext.TryGetSwitch suppressLoaderErrorsSwitch
+        switchFound && switchEnabled
+
+    let private newPrecompilerLoaderException =
+        let msg =
+            sprintf "Failed to load a precompiled grammar. Try rebuilding the assembly with the latest version \
+of Farkle. To suppress this error and build the grammar, set the '%s' AppContext switch to true."
+                suppressLoaderErrorsSwitch
+        fun (innerExn: exn) -> PrecompilerLoaderException(msg, innerExn)
+
     /// Gets the manifest resource name of the precompiled grammar with the given name.
     let getPrecompiledGrammarResourceName name = sprintf "%s.precompiled.egtn" name
 
@@ -55,17 +71,20 @@ module internal Loader =
     let getGrammarOrBuild (df: DesigntimeFarkle) =
         match df with
         | :? PrecompilableDesigntimeFarkle as pcdf ->
-            // The grammar is loaded from an EGTneo file in the
-            // assembly's resources, if it exists. Errors reading
-            // the grammar are unexpected and will throw.
-            let grammar =
-                pcdf.Name
-                |> getPrecompiledGrammarResourceName
-                |> pcdf.DeclaringAssembly.GetManifestResourceStream
-                |> function
-                | null -> DFB.buildGrammarOnly pcdf.Grammar
-                | stream -> stream |> EGT.ofStream |> Ok
-            grammar
+            try
+                let grammar =
+                    pcdf.Name
+                    |> getPrecompiledGrammarResourceName
+                    |> pcdf.DeclaringAssembly.GetManifestResourceStream
+                    |> function
+                    | null -> DFB.buildGrammarOnly pcdf.Grammar
+                    | stream -> stream |> EGT.ofStream |> Ok
+                grammar
+            with
+            | _ when shouldSuppressLoaderErrors ->
+                DFB.buildGrammarOnly pcdf.Grammar
+            | e ->
+                raise(newPrecompilerLoaderException e)
         | df -> df |> DFB.createGrammarDefinition |> DFB.buildGrammarOnly
 
     /// Marks the given designtime Farkle as eligible to